requirements.txt
langchain_community
tiktoken
langchain-openai
langchain-groq
langchainhub
chromadb
langchain
langchain-text-splitters
python-dotenv
pyyaml
pydantic
sentence-transformers
langchain_huggingface
fastapi